+++ src/storage/FileInfo.hxx
@@ -48,7 +48,7 @@ struct StorageFileInfo {
 	 * Device id and inode number.  0 means unknown / not
 	 * applicable.
 	 */
-	unsigned device, inode;
+	uint64_t device, inode;
 
 	constexpr bool IsRegular() const {
 		return type == Type::REGULAR;
